fix test case that broke in Django 2.2 (#21)
Django no longer always starts a transaction when a single query is being performed https://docs.djangoproject.com/en/2.2/releases/2.2/

This broke a test that counted SQL statements. I've adjusted the test do pick a number based on the Django version. We'll be able to delete it after we stop supporting Django 2.1