feat: split logging into create and update (#25)
When using this, a "create" might be unusual while "update" may not be (or vice-versa). If you want logging for just one of those, you can enable logging for `obj_update.create` or `obj_update.update`. Both events go to `obj_update` so you can continue to use that level too.